requirements.txt
# Core deep learning
torch>=1.13.0
torchvision>=0.14.0

# Image processing
Pillow>=9.0.0
scikit-image>=0.19.0
numpy>=1.21.0

# Visualisation
matplotlib>=3.5.0

# Web app
streamlit>=1.28.0
